requirements.txt
tavily-python
openai 
requests 
beautifulsoup4 
python-dotenv
groq